[#11] 매수, 매도 정보 추가 기능 구현
- companyName을 id로 매핑하는 작업 필요
