+++ src/frontend/src/lib/api/stockDetail.ts
@@ -1,0 +1,4 @@
+import axios from '../axios';
+
+export const getSingleStockDetail = (companyName: string) =>
+  axios.get('/stockdetail', { params: { companyName } });
